refactored to-prop file name
